add more to design; move all spacial things except positions to new src/hdr spacial; create src/hdr character for the main character; abstract a few usages of classes to make parameters more straight forward
