+++ design.md
@@ -12,7 +12,7 @@ small amount of credits earned over time, with killing enemies being the primary
 
 move <left|right|^up|^down>
 shoot [^direction] (or maybe auto shoot at interval)
-^get_nearest_enemy [^type]
+^get\_nearest\_enemy [^type]
 
 # ONE-TIME-UPGRADES (other than the indicated in commands)
 
@@ -26,8 +26,12 @@ firing speed -- how fast you shoot your primary weapon
 health -- how much damage you can take
 shields -- extra damage you can take, which regenerates if damage is not taken for awhile
 multishot -- shoots 3 projectiles instead of 1 every once in awhile
+script command speed -- decreases the time between script commands being run
+script branch speed -- decreases the time it takes for a branch in script to be run
+script syscall speed -- decreases the time it takes for a syscall in script to be run (things like obtaining data from a radar)
+movement speed -- decreases the time it takes to move from one grid cell to the next via move command
 
 # STARTING-OPTIONS
 
-starting difficulty and credit scale -- higher difficulty means higher credits
+starting difficulty and credit scale -- higher difficulty means higher credits from enemies, not (or really not much) from time passed
