+++ src/bouncing_pellet.hpp
@@ -7,7 +7,7 @@ class BouncingPellet : public MoveableEntity {
 public:
     BouncingPellet(float start_x, float start_y, float angle);
     void Draw(const DrawPosition &grid_tl_pos, float grid_cell_width, float grid_cell_height) override;
-    void Update() override;
+    void Update(int fps) override;
 };
 
 #endif // BOUNCING_PELLET_HPP
